+++ project/src/components/common/header/index.js
@@ -1,1 +1,1 @@
-export { default } from './Header';
+export { default } from './header';
