+++ project/src/components/common/loading-screen/index.js
@@ -1,1 +1,1 @@
-export * from './LoadingScreen';
+export * from './loading-screen';
