+++ project/src/components/common/loading-screen/loading-screen.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './LoadingScreen.css';
+import './loading-screen.css';
 
 export function LoadingScreen() {
   return (
