+++ project/src/components/common/logo/index.js
@@ -1,1 +1,1 @@
-export * from './Logo';
+export * from './logo';
