+++ project/src/components/common/map/index.js
@@ -1,1 +1,1 @@
-export * from './Map';
+export * from './map';
