+++ project/src/components/common/place-card-list/index.js
@@ -1,1 +1,1 @@
-export * from './PlaceCardList';
+export * from './place-card-list';
