+++ project/src/components/common/place-card-list/place-card/index.js
@@ -1,1 +1,1 @@
-export { default } from './PlaceCard';
+export { default } from './place-card';
