+++ project/src/components/common/place-card-list/place-card/place-card.jsx
