+++ project/src/components/common/private-route/index.js
@@ -1,1 +1,1 @@
-export {default} from './PrivateRoute';
+export {default} from './private-route';
