+++ project/src/components/pages/chosen/index.js
@@ -1,1 +1,1 @@
-export { default } from './Chosen';
+export { default } from './chosen';
