+++ project/src/components/pages/main/index.js
@@ -1,1 +1,1 @@
-export { default } from './Main';
+export { default } from './main';
