+++ project/src/components/pages/main/location-list/index.js
@@ -1,1 +1,1 @@
-export { default } from './LocationList';
+export { default } from './location-list';
