+++ project/src/components/pages/main/location-list/location-list.jsx
