+++ project/src/components/pages/main/sort-list/index.js
@@ -1,1 +1,1 @@
-export { default } from './SortList';
+export { default } from './sort-list';
