+++ project/src/components/pages/not-found/index.js
@@ -1,1 +1,1 @@
-export { default } from './NotFound';
+export { default } from './not-found';
