+++ project/src/components/pages/not-found/not-found.jsx
@@ -1,6 +1,6 @@
 import React from 'react-dom';
 import { Link } from 'react-router-dom';
-import './NotFound.css';
+import './not-found.css';
 
 function NotFound() {
   return (
