+++ project/src/components/pages/room/form/index.js
@@ -1,1 +1,1 @@
-export { default } from './Form';
+export { default } from './form';
