+++ project/src/components/pages/room/form/rating/index.js
@@ -1,1 +1,1 @@
-export { default } from './Rating';
+export { default } from './rating';
