+++ project/src/components/pages/room/form/rating/radio-button/index.js
@@ -1,1 +1,1 @@
-export { default } from './RadioButton';
+export { default } from './radio-button';
