+++ project/src/components/pages/room/reviews-list/index.js
@@ -1,1 +1,1 @@
-export { default } from './ReviewsList';
+export { default } from './reviews-list';
