+++ project/src/components/pages/room/reviews-list/reviews-item/index.js
@@ -1,1 +1,1 @@
-export { default } from './ReviewsItem';
+export { default } from './reviews-item';
