+++ project/src/components/pages/room/reviews-list/reviews-item/reviews-item.jsx
