+++ project/src/components/pages/sign-in/index.js
@@ -1,1 +1,1 @@
-export * from './SignIn';
+export * from './sign-in';
